requirements.txt
fake-bpy-module-3.6